chore: Move cards to components for reuse elsewhere
