+++ src/screens/SeriesScreen.tsx
@@ -3,8 +3,6 @@ import {
   View,
   Text,
   StyleSheet,
-  FlatList,
-  Image,
   ActivityIndicator,
 } from 'react-native';
 import { useXtream } from '../context/XtreamContext';
@@ -13,13 +11,10 @@ import { DrawerScreenPropsType } from '../navigation/types';
 import { XtreamCategory, XtreamSeries } from '../types/xtream';
 import { scaledPixels } from '../hooks/useScale';
 import { FocusablePressable } from '../components/FocusablePressable';
+import { SeriesCard } from '../components/SeriesCard';
 import { SpatialNavigationNode, SpatialNavigationVirtualizedGrid, SpatialNavigationVirtualizedList } from 'react-tv-space-navigation';
 
-// Card dimensions for consistent sizing (same as VOD)
-const CARD_WIDTH = scaledPixels(200);
-const CARD_MARGIN = scaledPixels(12);
-
-export function SeriesScreen({ navigation }: DrawerScreenPropsType<'Series'>) {
+export function SeriesScreen(_props: DrawerScreenPropsType<'Series'>) {
   const { isConfigured, seriesCategories, series, fetchSeries } = useXtream();
   const [selectedCategory, setSelectedCategory] = useState<string | undefined>(undefined);
   const [isLoading, setIsLoading] = useState(false);
@@ -61,35 +56,7 @@ export function SeriesScreen({ navigation }: DrawerScreenPropsType<'Series'>) {
   );
 
   const renderSeriesItem = ({ item }: { item: XtreamSeries }) => (
-    <FocusablePressable
-      style={({ isFocused }) => [
-        styles.seriesCard,
-        isFocused && styles.seriesCardFocused,
-      ]}
-      onSelect={() => {
-        // @ts-ignore
-        navigation.navigate('SeriesDetails', { item });
-      }}
-    >
-      <Image
-        source={{ uri: item.cover || 'https://via.placeholder.com/150x225' }}
-        style={styles.seriesPoster}
-        resizeMode="cover"
-      />
-      <View style={styles.seriesInfo}>
-        <Text style={styles.seriesName} numberOfLines={1}>
-          {item.name}
-        </Text>
-        <View style={styles.seriesMeta}>
-          <Text style={styles.seriesRating}>★ {item.rating || 'N/A'}</Text>
-          {(item.release_date || item.releaseDate) && (
-            <Text style={styles.seriesYear}>
-              {(item.release_date || item.releaseDate)?.substring(0, 4)}
-            </Text>
-          )}
-        </View>
-      </View>
-    </FocusablePressable>
+    <SeriesCard item={item} />
   );
 
   if (!isConfigured) {
@@ -218,50 +185,5 @@ const styles = StyleSheet.create({
   },
   seriesGrid: {
     padding: scaledPixels(20),
-  },
-  seriesCard: {
-    width: CARD_WIDTH,
-    margin: CARD_MARGIN,
-    backgroundColor: colors.card,
-    borderRadius: scaledPixels(12),
-    overflow: 'hidden',
-    borderWidth: 3,
-    borderColor: 'transparent',
-  },
-  seriesCardFocused: {
-    borderColor: colors.primary,
-    transform: [{ scale: 1.08 }],
-    zIndex: 10,
-    shadowColor: colors.primary,
-    shadowOffset: { width: 0, height: 0 },
-    shadowOpacity: 0.6,
-    shadowRadius: 15,
-    elevation: 10,
-  },
-  seriesPoster: {
-    width: '100%',
-    aspectRatio: 2 / 3,
-  },
-  seriesInfo: {
-    padding: scaledPixels(12),
-  },
-  seriesName: {
-    color: colors.text,
-    fontSize: scaledPixels(16),
-    fontWeight: '500',
-  },
-  seriesMeta: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: scaledPixels(8),
-    marginTop: scaledPixels(4),
-  },
-  seriesRating: {
-    color: colors.warning,
-    fontSize: scaledPixels(14),
-  },
-  seriesYear: {
-    color: colors.textSecondary,
-    fontSize: scaledPixels(14),
   },
 });
