+++ src/screens/VODScreen.tsx
@@ -3,8 +3,6 @@ import {
   View,
   Text,
   StyleSheet,
-  FlatList,
-  Image,
   ActivityIndicator,
 } from 'react-native';
 import { useXtream } from '../context/XtreamContext';
@@ -13,13 +11,10 @@ import { DrawerScreenPropsType } from '../navigation/types';
 import { XtreamCategory, XtreamVodStream } from '../types/xtream';
 import { scaledPixels } from '../hooks/useScale';
 import { FocusablePressable } from '../components/FocusablePressable';
+import { MovieCard } from '../components/MovieCard';
 import { SpatialNavigationNode, SpatialNavigationVirtualizedGrid, SpatialNavigationVirtualizedList } from 'react-tv-space-navigation';
 
-// Card dimensions for consistent sizing
-const CARD_WIDTH = scaledPixels(200);
-const CARD_MARGIN = scaledPixels(12);
-
-export function VODScreen({ navigation }: DrawerScreenPropsType<'VOD'>) {
+export function VODScreen(_props: DrawerScreenPropsType<'VOD'>) {
   const { isConfigured, vodCategories, vodStreams, fetchVodStreams } = useXtream();
   const [selectedCategory, setSelectedCategory] = useState<string | undefined>(undefined);
   const [isLoading, setIsLoading] = useState(false);
@@ -61,28 +56,7 @@ export function VODScreen({ navigation }: DrawerScreenPropsType<'VOD'>) {
   );
 
   const renderMovieItem = ({ item }: { item: XtreamVodStream }) => (
-    <FocusablePressable
-      style={({ isFocused }) => [
-        styles.movieCard,
-        isFocused && styles.movieCardFocused,
-      ]}
-      onSelect={() => {
-        // @ts-ignore
-        navigation.navigate('Details', { item });
-      }}
-    >
-      <Image
-        source={{ uri: item.stream_icon || 'https://via.placeholder.com/150x225' }}
-        style={styles.moviePoster}
-        resizeMode="cover"
-      />
-      <View style={styles.movieInfo}>
-        <Text style={styles.movieName} numberOfLines={1}>
-          {item.name}
-        </Text>
-        <Text style={styles.movieRating}>★ {item.rating || 'N/A'}</Text>
-      </View>
-    </FocusablePressable>
+    <MovieCard item={item} />
   );
 
   if (!isConfigured) {
@@ -211,41 +185,5 @@ const styles = StyleSheet.create({
   },
   movieGrid: {
     padding: scaledPixels(20),
-  },
-  movieCard: {
-    width: CARD_WIDTH,
-    margin: CARD_MARGIN,
-    backgroundColor: colors.card,
-    borderRadius: scaledPixels(12),
-    overflow: 'hidden',
-    borderWidth: 3,
-    borderColor: 'transparent',
-  },
-  movieCardFocused: {
-    borderColor: colors.primary,
-    transform: [{ scale: 1.08 }],
-    zIndex: 10,
-    shadowColor: colors.primary,
-    shadowOffset: { width: 0, height: 0 },
-    shadowOpacity: 0.6,
-    shadowRadius: 15,
-    elevation: 10,
-  },
-  moviePoster: {
-    width: '100%',
-    aspectRatio: 2 / 3,
-  },
-  movieInfo: {
-    padding: scaledPixels(12),
-  },
-  movieName: {
-    color: colors.text,
-    fontSize: scaledPixels(16),
-    fontWeight: '500',
-  },
-  movieRating: {
-    color: colors.warning,
-    fontSize: scaledPixels(14),
-    marginTop: scaledPixels(4),
   },
 });
